Fixed extract_command manager test
