+++ scale/messaging/management/commands/scale_echo_message.py
@@ -8,7 +8,7 @@ import logging
 from datetime import datetime
 
 from django.core.management.base import BaseCommand
-from messaging.messages import CommandMessageManager
+from messaging.manager import CommandMessageManager
 from messaging.messages.echo import EchoCommandMessage
 
 logger = logging.getLogger(__name__)
