+++ scale/messaging/management/commands/scale_process_messages.py
@@ -8,7 +8,7 @@ import logging
 import signal
 
 from django.core.management.base import BaseCommand
-from messaging.messages import CommandMessageManager
+from messaging.manager import CommandMessageManager
 
 logger = logging.getLogger(__name__)
