+++ scale/messaging/manager.py
